+++ src/parsers/ConfigurationParser.py
@@ -9,7 +9,8 @@ from src.utils import parse_value_with_default
 from src.utils import check_linux_permissions
 from src.utils import parse_logging_level
 from src.utils import match_port_to_protocol
-from src.data import default_data
+from src.data import default
+from src.data.validator import validate_ip, validate_domain, validate_network_port, validate_http_response_code
 
 
 class ConfigParser:
@@ -101,6 +102,24 @@ class ConfigParser:
 
         return False
 
+    def validate_dns(self, domain: str, primary_answer: str, failover_answers: list) -> bool:
+        """
+        Check if domain is valid domain, check if primary_answers and failover answers are valid ip addresses
+        :param domain:
+        :param primary_answer:
+        :param failover_answers:
+        :return: True if all everything is valid, False if not
+        """
+        if validate_domain(domain=domain) is False:
+            return False
+        elif validate_ip(ip=primary_answer) is False:
+            return False
+        else:
+            for host in failover_answers:
+                if validate_ip(host) is False:
+                    return False
+            return True
+
     def parse_http(self):
         """
         Parse http jobs, create dictionary compatible with TestHosts.py
@@ -108,39 +127,55 @@ class ConfigParser:
         """
         job_index = 0
         for jobs in self.file_content['http_jobs']:
+
             try:
                 job = jobs['job']
                 self.http_configs[job_index] = {}
-                self.http_configs[job_index]['dns_domain'] = job['domain']
-                self.http_configs[job_index]['dns_answer'] = job['answers']['primary']
+                dns_domain = job['domain']
+                dns_answer = job['answers']['primary']
 
                 if 'failover' in job['answers']:
                     if job['answers']['failover'] is None:
-                        self.http_configs[job_index]['dns_answer_failover'] = []
+                        dns_failover = []
                     else:
-                        self.http_configs[job_index]['dns_answer_failover'] = job['answers']['failover']
+                        dns_failover = job['answers']['failover']
                 else:
-                    self.http_configs[job_index]['dns_answer_failover'] = []
+                    dns_failover = []
 
-                self.http_configs[job_index]['interval'] = parse_value_with_default(content=job, key='interval',
-                                                                                    default_value=
-                                                                                    default_data.HttpJob.interval)
-                self.http_configs[job_index]['status_code'] = parse_value_with_default(content=job, key='status',
-                                                                                       default_value=
-                                                                                       default_data.HttpJob.status)
-                self.http_configs[job_index]['proto'] = parse_value_with_default(content=job, key='proto',
-                                                                                 default_value='http')
-                self.http_configs[job_index]['port'] = parse_value_with_default(content=job, key='port',
-                                                                                default_value=match_port_to_protocol(
-                                                                                    proto=self.http_configs[job_index][
-                                                                                        'proto'],
-                                                                                    default_port=
-                                                                                    default_data.HttpJob.port)
-                                                                                )
+                interval = parse_value_with_default(content=job, key='interval',
+                                                    default_value=default.HttpJob.interval)
 
-                self.http_configs[job_index]['timeout'] = parse_value_with_default(content=job, key='timeout',
-                                                                                   default_value=
-                                                                                   default_data.HttpJob.timeout)
+                status_code = parse_value_with_default(content=job, key='status',
+                                                       default_value=default.HttpJob.status)
+
+                proto = parse_value_with_default(content=job, key='proto',
+                                                 default_value=default.HttpJob.proto)
+
+                port = parse_value_with_default(content=job, key='port',
+                                                default_value=match_port_to_protocol(
+                                                    proto=proto, default_port=default.HttpJob.port))
+
+                timeout = parse_value_with_default(content=job, key='timeout',
+                                                   default_value=default.HttpJob.timeout)
+
+            except KeyError:
+                logging.error("Error in config file, http_jobs KeyError")
+                break
+
+            data_valid = self.validate_dns(domain=dns_domain, primary_answer=dns_answer, failover_answers=dns_failover)
+            data_valid = data_valid and validate_network_port(port=port)
+            data_valid = data_valid and validate_http_response_code(code=status_code)
+            if data_valid:
+
+                self.http_configs[job_index] = {}
+                self.http_configs[job_index]['dns_domain'] = dns_domain
+                self.http_configs[job_index]['dns_answer'] = dns_answer
+                self.http_configs[job_index]['dns_answer_failover'] = dns_failover
+                self.http_configs[job_index]['interval'] = interval
+                self.http_configs[job_index]['status_code'] = status_code
+                self.http_configs[job_index]['proto'] = proto
+                self.http_configs[job_index]['port'] = port
+                self.http_configs[job_index]['timeout'] = timeout
 
                 logging.debug(msg="http-domain " + self.http_configs[job_index]['dns_domain'])
                 logging.debug(msg="http-interval " + str(self.http_configs[job_index]['interval']))
@@ -152,8 +187,9 @@ class ConfigParser:
                 logging.debug(msg="http-timeout " + str(self.http_configs[job_index]['timeout']))
 
                 job_index = job_index + 1
-            except KeyError:
-                logging.error("Error in config file, http_jobs KeyError")
+
+            else:
+                logging.info("Job for domain: " + dns_domain + " not added, due to invalid parameters")
 
     def parse_ping(self):
         """
@@ -161,31 +197,40 @@ class ConfigParser:
         :return:
         """
         job_index = 0
-
         for jobs in self.file_content['ping_jobs']:
             try:
                 job = jobs['job']
-                self.ping_configs[job_index] = {}
-                self.ping_configs[job_index]['dns_domain'] = job['domain']
-                self.ping_configs[job_index]['dns_answer'] = job['answers']['primary']
+                dns_domain = job['domain']
+                dns_answer = job['answers']['primary']
 
                 if 'failover' in job['answers']:
                     if job['answers']['failover'] is None:
-                        self.ping_configs[job_index]['dns_answer_failover'] = []
+                        dns_failover = []
                     else:
-                        self.ping_configs[job_index]['dns_answer_failover'] = job['answers']['failover']
+                        dns_failover = job['answers']['failover']
                 else:
-                    self.ping_configs[job_index]['dns_answer_failover'] = []
+                    dns_failover = []
 
-                self.ping_configs[job_index]['interval'] = parse_value_with_default(content=job, key='interval',
-                                                                                    default_value=
-                                                                                    default_data.PingJob.interval)
-                self.ping_configs[job_index]['timeout'] = parse_value_with_default(content=job, key='timeout',
-                                                                                   default_value=
-                                                                                   default_data.PingJob.timeout)
-                self.ping_configs[job_index]['count'] = parse_value_with_default(content=job, key='count',
-                                                                                 default_value=
-                                                                                 default_data.PingJob.count)
+                interval = parse_value_with_default(content=job, key='interval',
+                                                    default_value=default.PingJob.interval)
+                timeout = parse_value_with_default(content=job, key='timeout',
+                                                   default_value=default.PingJob.timeout)
+                count = parse_value_with_default(content=job, key='count',
+                                                 default_value=default.PingJob.count)
+
+            except KeyError:
+                logging.error("Error in config file, ping_jobs KeyError")
+
+            data_valid = self.validate_dns(domain=dns_domain, primary_answer=dns_answer, failover_answers=dns_failover)
+
+            if data_valid:
+                self.ping_configs[job_index] = {}
+                self.ping_configs[job_index]['dns_domain'] = dns_domain
+                self.ping_configs[job_index]['dns_answer'] = dns_answer
+                self.ping_configs[job_index]['dns_answer_failover'] = dns_failover
+                self.ping_configs[job_index]['interval'] = interval
+                self.ping_configs[job_index]['timeout'] = timeout
+                self.ping_configs[job_index]['count'] = count
 
                 logging.debug(msg="ping-domain " + self.ping_configs[job_index]['dns_domain'])
                 logging.debug(msg="ping-interval " + str(self.ping_configs[job_index]['interval']))
@@ -195,29 +240,40 @@ class ConfigParser:
                 logging.debug(msg="ping-failover " + ' '.join(self.ping_configs[job_index]['dns_answer_failover']))
 
                 job_index = job_index + 1
-            except KeyError:
-                logging.error("Error in config file, ping_jobs KeyError")
+
+            else:
+                logging.info("Job for domain: " + dns_domain + " not added, due to invalid parameters")
 
     def parser_static_entry(self):
         job_index = 0
         for jobs in self.file_content['static_entry']:
             try:
                 job = jobs['job']
+
+                domain = job['domain']
+                answer = job['answer']
+                interval = parse_value_with_default(content=job, key='interval',
+                                                    default_value=default.StaticEntry.interval)
+
+            except KeyError:
+                logging.error("Error in config file, static_entry KeyError")
+
+            data_valid = self.validate_dns(domain=domain, primary_answer=answer, failover_answers=[])
+
+            if data_valid:
                 self.static_entry_configs[job_index] = {}
-                self.static_entry_configs[job_index]['domain'] = job['domain']
-                self.static_entry_configs[job_index]['answer'] = job['answer']
-                self.static_entry_configs[job_index]['interval'] = parse_value_with_default(
-                                                                                    content=job,
-                                                                                    key='interval',
-                                                                                    default_value=
-                                                                                    default_data.StaticEntry.interval)
+                self.static_entry_configs[job_index]['domain'] = domain
+                self.static_entry_configs[job_index]['answer'] = answer
+                self.static_entry_configs[job_index]['interval'] = interval
+
                 logging.debug(msg="data-entry-domain " + self.static_entry_configs[job_index]['domain'])
                 logging.debug(msg="data-entry-answer " + self.static_entry_configs[job_index]['answer'])
                 logging.debug(msg="data-entry-interval " + str(self.static_entry_configs[job_index]['interval']))
 
                 job_index += 1
-            except KeyError:
-                logging.error("Error in config file, static_entry KeyError")
+
+            else:
+                logging.info("Job for domain: " + domain + " not added, due to invalid parameters")
 
     def parse_api(self):
         """
@@ -225,36 +281,53 @@ class ConfigParser:
         :return:
         """
         try:
-            self.api_config['host'] = self.file_content['api']['host']
-            self.api_config['username'] = self.file_content['api']['username']
-            self.api_config['passwd'] = self.file_content['api']['passwd']
+            host = self.file_content['api']['host']
+            username = self.file_content['api']['username']
+            passwd = self.file_content['api']['passwd']
 
-            self.api_config['proto'] = parse_value_with_default(content=self.file_content['api'], key='proto',
-                                                                default_value=default_data.Api.proto)
-            self.api_config['port'] = parse_value_with_default(content=self.file_content['api'], key='port',
-                                                               default_value=default_data.Api.port)
-            self.api_config['timeout'] = parse_value_with_default(content=self.file_content['api'], key='timeout',
-                                                                  default_value=default_data.Api.timeout)
+            proto = parse_value_with_default(content=self.file_content['api'], key='proto',
+                                                                default_value=default.Api.proto)
+            port = parse_value_with_default(content=self.file_content['api'], key='port',
+                                                               default_value=default.Api.port)
+            timeout = parse_value_with_default(content=self.file_content['api'], key='timeout',
+                                                                  default_value=default.Api.timeout)
             if 'startup' in self.file_content['api']:
-                self.api_config['startup'] = {}
-                self.api_config['startup']['test'] = parse_value_with_default(
+                startup_test = parse_value_with_default(
                     content=self.file_content['api']['startup'],
-                    key='test', default_value=default_data.Api.Startup.test)
-                self.api_config['startup']['timeout'] = parse_value_with_default(
+                    key='test', default_value=default.Api.Startup.test)
+                startup_timeout = parse_value_with_default(
                     content=self.file_content['api']['startup'],
-                    key='timeout', default_value=default_data.Api.Startup.timeout)
-                self.api_config['startup']['exit_on_fail'] = parse_value_with_default(
+                    key='timeout', default_value=default.Api.Startup.timeout)
+                startup_exit_on_fall = parse_value_with_default(
                     content=self.file_content['api']['startup'], key='exit_on_fail',
-                    default_value=default_data.Api.Startup.exit_on_false)
-                self.api_config['startup']['retry_after'] = parse_value_with_default(
+                    default_value=default.Api.Startup.exit_on_false)
+                startup_retry_after = parse_value_with_default(
                     content=self.file_content['api']['startup'], key='retry_after',
-                    default_value=default_data.Api.Startup.retry_after)
+                    default_value=default.Api.Startup.retry_after)
             else:
-                self.api_config['startup'] = {}
-                self.api_config['startup']['test'] = True
-                self.api_config['startup']['timeout'] = 10
-                self.api_config['startup']['exit_on_fail'] = False
-                self.api_config['startup']['retry_after'] = 10
+                startup_test = True
+                startup_timeout = 10
+                startup_exit_on_fall = False
+                startup_retry_after = 10
+
+        except KeyError:
+            logging.error("Config file error / api / KeyError")
+            exit(-2)
+
+        data_valid = validate_ip(ip=host) or validate_domain(domain=host)
+        data_valid = data_valid and validate_network_port(port=port)
+        if data_valid:
+            self.api_config['host'] = host
+            self.api_config['username'] = username
+            self.api_config['passwd'] = passwd
+            self.api_config['proto'] = proto
+            self.api_config['port'] = port
+            self.api_config['timeout'] = timeout
+            self.api_config['startup'] = {}
+            self.api_config['startup']['test'] = startup_test
+            self.api_config['startup']['timeout'] = startup_timeout
+            self.api_config['startup']['exit_on_fail'] = startup_exit_on_fall
+            self.api_config['startup']['retry_after'] = startup_retry_after
 
             logging.debug(msg="api-host " + self.api_config['host'])
             logging.debug(msg="api-username " + self.api_config['username'])
@@ -266,10 +339,8 @@ class ConfigParser:
             logging.debug(msg='api-startup-timeout ' + str(self.api_config['startup']['timeout']))
             logging.debug(msg='api-startup-exit_on_fail ' + str(self.api_config['startup']['exit_on_fail']))
             logging.debug(msg='api-startup-test-retry_after ' + str(self.api_config['startup']['retry_after']))
-
-        except KeyError:
-            logging.error("Config file error / api / KeyError")
-            exit(-2)
+        else:
+            logging.info("Api configuration error")
 
     def parse_config(self):
         """
@@ -279,24 +350,24 @@ class ConfigParser:
         try:
             if 'config' in self.file_content:
                 self.config_config['wait'] = parse_value_with_default(
-                    content=self.file_content['config'], key='wait', default_value=default_data.Config.wait)
+                    content=self.file_content['config'], key='wait', default_value=default.Config.wait)
                 log_level = parse_value_with_default(
                     content=self.file_content['config'], key="log_level", default_value="N/A")
 
                 self.config_config['log_level'] = parse_logging_level(logging_str=log_level)
 
                 self.config_config['log_file'] = parse_value_with_default(
-                    content=self.file_content['config'], key='log_file', default_value=default_data.Config.log_file)
+                    content=self.file_content['config'], key='log_file', default_value=default.Config.log_file)
 
                 self.config_config['entry_exist'] = parse_value_with_default(content=self.file_content['config'],
                                                                              key='entry_exist',
                                                                              default_value=
-                                                                             default_data.Config.entry_exist)
+                                                                             default.Config.entry_exist)
             else:
-                self.config_config['wait'] = default_data.Config.wait
-                self.config_config['log_level'] = default_data.Config.log_level
-                self.config_config['log_file'] = default_data.Config.log_file
-                self.config_config['entry_exist'] = default_data.Config.entry_exist
+                self.config_config['wait'] = default.Config.wait
+                self.config_config['log_level'] = default.Config.log_level
+                self.config_config['log_file'] = default.Config.log_file
+                self.config_config['entry_exist'] = default.Config.entry_exist
 
             logging.debug(msg="config-wait " + str(self.config_config['wait']))
             logging.debug(msg="config-log_level " + str(self.config_config['log_level']))